+++ mem_display.c
@@ -2,24 +2,26 @@
 
 
 #include "command_interface.h"
-int MemDisplayValidate(char* params);
-int MemDisplayExecute(char* params);
+#include "stdint.h"
+
+uint8_t MemDisplayValidate(char* params);
+uint8_t MemDisplayExecute(char* params);
 
 COMMAND_INTERFACE_STRUCT MemDisplayCommandInterface =
 {
-	"MDIS",
-	"Displays a block of memory. Takes 2 arguments: Starting Address and Size (in 4 byte words). Ex: MDIS 0x12345678 100",
-	MemDisplayValidate,
-	MemDisplayExecute,
+  "MDIS",
+  "Displays a block of memory. Takes 2 arguments: Starting Address and Size (in 4 byte words). Ex: MDIS 0x12345678 100",
+  MemDisplayValidate,
+  MemDisplayExecute,
 };
 
 
-int MemDisplayValidate(char* params)
+uint8_t MemDisplayValidate(char* params)
 {
-	return 1;
+  return 1;
 }
 
-int MemDisplayExecute(char* params)
+uint8_t MemDisplayExecute(char* params)
 {
-	return 1;
+  return 1;
 }
