+++ mem_ver_pattern.c
@@ -1,23 +1,25 @@
 
 #include "command_interface.h"
-int MemVerPatternValidate(char* params);
-int MemVerPatternExecute(char* params);
+#include "stdint.h"
+
+uint8_t MemVerPatternValidate(char* params);
+uint8_t MemVerPatternExecute(char* params);
 
 COMMAND_INTERFACE_STRUCT MemVerPatternCommandInterface =
 {
-	"MVPA",
-	"Verifies Pattern matches the memory space designated by address and size. Ex MVPA 0x12345678 100",
-	MemVerPatternValidate,
-	MemVerPatternExecute,
+  "MVPA",
+  "Verifies Pattern matches the memory space designated by address and size. Ex MVPA 0x12345678 100",
+  MemVerPatternValidate,
+  MemVerPatternExecute,
 };
 
 
-int MemVerPatternValidate(char* params)
+uint8_t MemVerPatternValidate(char* params)
 {
-	return 1;
+  return 1;
 }
 
-int MemVerPatternExecute(char* params)
+uint8_t MemVerPatternExecute(char* params)
 {
-	return 1;
+  return 1;
 }
